+++ internal/component/otelcol/config_queue.go
@@ -15,7 +15,6 @@ type QueueArguments struct {
 	Enabled         bool   `alloy:"enabled,attr,optional"`
 	NumConsumers    int    `alloy:"num_consumers,attr,optional"`
 	QueueSize       int64  `alloy:"queue_size,attr,optional"`
-	Blocking        bool   `alloy:"blocking,attr,optional"`
 	BlockOnOverflow bool   `alloy:"block_on_overflow,attr,optional"`
 	Sizer           string `alloy:"sizer,attr,optional"`
 
@@ -80,7 +79,6 @@ func (args *QueueArguments) Convert() (*otelexporterhelper.QueueBatchConfig, err
 		Enabled:         args.Enabled,
 		NumConsumers:    args.NumConsumers,
 		QueueSize:       args.QueueSize,
-		Blocking:        args.Blocking,
 		BlockOnOverflow: args.BlockOnOverflow,
 		Sizer:           *sizer,
 	}